Add compatibility with native Wifi
The wifi manager instance is only used to call `get`, `post` and `put`, so it could be substituted with requests, or an instance of adafruit_requests.Session for ESP32S2. All it needs is not to test the type of the passed object. I don't have a any Hue device to test.